Move realtime lowering into a separate pass
