+++ cudaq/test/Transforms/device_call_realtime_device_id.qke
@@ -6,24 +6,30 @@
 // the terms of the Apache License 2.0 which accompanies this distribution.   //
 // ========================================================================== //
 
-// RUN: cudaq-opt -pass-pipeline='builtin.module(canonicalize,distributed-device-call{realtime=1},cse,canonicalize)' %s | FileCheck %s
+// RUN: cudaq-opt -pass-pipeline='builtin.module(canonicalize,realtime-device-call,cse,canonicalize)' %s | FileCheck %s
+
+func.func @kernel(%device : i32, %arg0 : i32, %arg1 : i32) -> i32 attributes {"cudaq-entrypoint", "cudaq-kernel", no_this} {
+  %0 = cc.device_call @addThem on %device(%arg0, %arg1) : (i32, i32, i32) -> i32
+  return %0 : i32
+}
 
 // CHECK-LABEL:   func.func @kernel(
 // CHECK-SAME:                      %[[ARG0:[0-9]+|[a-zA-Z$._-][a-zA-Z0-9$._-]*]]: i32,
 // CHECK-SAME:                      %[[ARG1:[0-9]+|[a-zA-Z$._-][a-zA-Z0-9$._-]*]]: i32,
 // CHECK-SAME:                      %[[ARG2:[0-9]+|[a-zA-Z$._-][a-zA-Z0-9$._-]*]]: i32) -> i32 attributes {"cudaq-entrypoint", "cudaq-kernel", no_this} {
-// CHECK:           %[[CONSTANT_0:.*]] = arith.constant 0 : i32
-// CHECK:           %[[CONSTANT_1:.*]] = arith.constant 0 : i64
-// CHECK:           %[[CONSTANT_2:.*]] = arith.constant 1517064222 : i32
-// CHECK:           %[[CONSTANT_3:.*]] = arith.constant 4 : i64
-// CHECK:           %[[CONSTANT_4:.*]] = arith.constant 8 : i64
+// CHECK:           %[[CONSTANT_0:.*]] = arith.constant 6 : i64
+// CHECK:           %[[CONSTANT_1:.*]] = arith.constant 0 : i32
+// CHECK:           %[[CONSTANT_2:.*]] = arith.constant 0 : i64
+// CHECK:           %[[CONSTANT_3:.*]] = arith.constant 8 : i64
+// CHECK:           %[[CONSTANT_4:.*]] = arith.constant 4 : i64
+// CHECK:           %[[CONSTANT_5:.*]] = arith.constant 1517064222 : i32
 // CHECK:           %[[ALLOCA_0:.*]] = cc.alloca !cc.ptr<i8>
 // CHECK:           %[[ALLOCA_1:.*]] = cc.alloca !cc.ptr<i8>
 // CHECK:           %[[ALLOCA_2:.*]] = cc.alloca !cc.ptr<i8>
 // CHECK:           %[[ALLOCA_3:.*]] = cc.alloca i64
-// CHECK:           cc.store %[[CONSTANT_1]], %[[ALLOCA_3]] : !cc.ptr<i64>
-// CHECK:           %[[VAL_0:.*]] = call @__cudaq_device_call_acquire_realtime_frame(%[[ARG0]], %[[CONSTANT_2]], %[[CONSTANT_4]], %[[CONSTANT_3]], %[[ALLOCA_0]], %[[ALLOCA_1]], %[[ALLOCA_2]]) : (i32, i32, i64, i64, !cc.ptr<!cc.ptr<i8>>, !cc.ptr<!cc.ptr<i8>>, !cc.ptr<!cc.ptr<i8>>) -> i32
-// CHECK:           %[[CMPI_0:.*]] = arith.cmpi ne, %[[VAL_0]], %[[CONSTANT_0]] : i32
+// CHECK:           cc.store %[[CONSTANT_2]], %[[ALLOCA_3]] : !cc.ptr<i64>
+// CHECK:           %[[VAL_0:.*]] = call @__cudaq_device_call_acquire_realtime_frame(%[[ARG0]], %[[CONSTANT_5]], %[[CONSTANT_3]], %[[CONSTANT_4]], %[[ALLOCA_0]], %[[ALLOCA_1]], %[[ALLOCA_2]]) : (i32, i32, i64, i64, !cc.ptr<!cc.ptr<i8>>, !cc.ptr<!cc.ptr<i8>>, !cc.ptr<!cc.ptr<i8>>) -> i32
+// CHECK:           %[[CMPI_0:.*]] = arith.cmpi ne, %[[VAL_0]], %[[CONSTANT_1]] : i32
 // CHECK:           cc.if(%[[CMPI_0]]) {
 // CHECK:             %[[CAST_0:.*]] = cc.cast signed %[[VAL_0]] : (i32) -> i64
 // CHECK:             func.call @__quantum__qis__trap(%[[CAST_0]]) : (i64) -> ()
@@ -38,25 +44,28 @@
 // CHECK:           %[[CAST_3:.*]] = cc.cast %[[COMPUTE_PTR_0]] : (!cc.ptr<i8>) -> !cc.ptr<i32>
 // CHECK:           cc.store %[[ARG2]], %[[CAST_3]] : !cc.ptr<i32>
 // CHECK:           %[[VAL_1:.*]] = call @__cudaq_device_call_dispatch_realtime_frame(%[[LOAD_0]], %[[ALLOCA_3]]) : (!cc.ptr<i8>, !cc.ptr<i64>) -> i32
-// CHECK:           %[[CMPI_1:.*]] = arith.cmpi ne, %[[VAL_1]], %[[CONSTANT_0]] : i32
+// CHECK:           %[[CMPI_1:.*]] = arith.cmpi ne, %[[VAL_1]], %[[CONSTANT_1]] : i32
 // CHECK:           cc.if(%[[CMPI_1]]) {
 // CHECK:             func.call @__cudaq_device_call_safely_release_realtime_frame(%[[LOAD_0]]) : (!cc.ptr<i8>) -> ()
 // CHECK:             %[[CAST_4:.*]] = cc.cast signed %[[VAL_1]] : (i32) -> i64
 // CHECK:             func.call @__quantum__qis__trap(%[[CAST_4]]) : (i64) -> ()
 // CHECK:           }
+// CHECK:           %[[LOAD_3:.*]] = cc.load %[[ALLOCA_3]] : !cc.ptr<i64>
+// CHECK:           %[[CMPI_2:.*]] = arith.cmpi ne, %[[LOAD_3]], %[[CONSTANT_4]] : i64
+// CHECK:           cc.if(%[[CMPI_2]]) {
+// CHECK:             func.call @__cudaq_device_call_safely_release_realtime_frame(%[[LOAD_0]]) : (!cc.ptr<i8>) -> ()
+// CHECK:             func.call @__quantum__qis__trap(%[[CONSTANT_0]]) : (i64) -> ()
+// CHECK:           }
 // CHECK:           %[[CAST_5:.*]] = cc.cast %[[LOAD_2]] : (!cc.ptr<i8>) -> !cc.ptr<i32>
-// CHECK:           %[[LOAD_3:.*]] = cc.load %[[CAST_5]] : !cc.ptr<i32>
+// CHECK:           %[[LOAD_4:.*]] = cc.load %[[CAST_5]] : !cc.ptr<i32>
 // CHECK:           call @__cudaq_device_call_safely_release_realtime_frame(%[[LOAD_0]]) : (!cc.ptr<i8>) -> ()
-// CHECK:           return %[[LOAD_3]] : i32
+// CHECK:           return %[[LOAD_4]] : i32
 // CHECK:         }
 // CHECK:         func.func private @addThem(i32, i32) -> i32 attributes {"cudaq-devicecall"}
 // CHECK:         func.func private @__cudaq_device_call_acquire_realtime_frame(i32, i32, i64, i64, !cc.ptr<!cc.ptr<i8>>, !cc.ptr<!cc.ptr<i8>>, !cc.ptr<!cc.ptr<i8>>) -> i32
 // CHECK:         func.func private @__cudaq_device_call_dispatch_realtime_frame(!cc.ptr<i8>, !cc.ptr<i64>) -> i32
 // CHECK:         func.func private @__cudaq_device_call_safely_release_realtime_frame(!cc.ptr<i8>)
+// CHECK:         func.func private @llvm.memcpy.p0.p0.i64(!cc.ptr<i8>, !cc.ptr<i8>, i64, i1)
 // CHECK:         func.func private @__quantum__qis__trap(i64)
-func.func @kernel(%device : i32, %arg0 : i32, %arg1 : i32) -> i32 attributes {"cudaq-entrypoint", "cudaq-kernel", no_this} {
-  %0 = cc.device_call @addThem on %device(%arg0, %arg1) : (i32, i32, i32) -> i32
-  return %0 : i32
-}
 
 func.func private @addThem(i32, i32) -> i32 attributes {"cudaq-devicecall"}
