+++ cudaq/test/Transforms/device_call_realtime_void.qke
@@ -6,21 +6,26 @@
 // the terms of the Apache License 2.0 which accompanies this distribution.   //
 // ========================================================================== //
 
-// RUN: cudaq-opt -pass-pipeline='builtin.module(canonicalize,distributed-device-call{realtime=1},cse,canonicalize)' %s | FileCheck %s
+// RUN: cudaq-opt -pass-pipeline='builtin.module(canonicalize,realtime-device-call,cse,canonicalize)' %s | FileCheck %s
+
+func.func @kernel(%arg0 : i32) attributes {"cudaq-entrypoint", "cudaq-kernel", no_this} {
+  cc.device_call @fireAndForget(%arg0) : (i32) -> ()
+  return
+}
 
 // CHECK-LABEL:   func.func @kernel(
 // CHECK-SAME:                      %[[ARG0:[0-9]+|[a-zA-Z$._-][a-zA-Z0-9$._-]*]]: i32) attributes {"cudaq-entrypoint", "cudaq-kernel", no_this} {
-// CHECK:           %[[CONSTANT_0:.*]] = arith.constant 0 : i32
-// CHECK:           %[[CONSTANT_1:.*]] = arith.constant 962244963 : i32
-// CHECK:           %[[CONSTANT_2:.*]] = arith.constant 0 : i64
-// CHECK:           %[[CONSTANT_3:.*]] = arith.constant 4 : i64
+// CHECK:           %[[CONSTANT_0:.*]] = arith.constant 4 : i64
+// CHECK:           %[[CONSTANT_1:.*]] = arith.constant 0 : i64
+// CHECK:           %[[CONSTANT_2:.*]] = arith.constant 962244963 : i32
+// CHECK:           %[[CONSTANT_3:.*]] = arith.constant 0 : i32
 // CHECK:           %[[ALLOCA_0:.*]] = cc.alloca !cc.ptr<i8>
 // CHECK:           %[[ALLOCA_1:.*]] = cc.alloca !cc.ptr<i8>
 // CHECK:           %[[ALLOCA_2:.*]] = cc.alloca !cc.ptr<i8>
 // CHECK:           %[[ALLOCA_3:.*]] = cc.alloca i64
-// CHECK:           cc.store %[[CONSTANT_2]], %[[ALLOCA_3]] : !cc.ptr<i64>
-// CHECK:           %[[VAL_0:.*]] = call @__cudaq_device_call_acquire_realtime_frame(%[[CONSTANT_0]], %[[CONSTANT_1]], %[[CONSTANT_3]], %[[CONSTANT_2]], %[[ALLOCA_0]], %[[ALLOCA_1]], %[[ALLOCA_2]]) : (i32, i32, i64, i64, !cc.ptr<!cc.ptr<i8>>, !cc.ptr<!cc.ptr<i8>>, !cc.ptr<!cc.ptr<i8>>) -> i32
-// CHECK:           %[[CMPI_0:.*]] = arith.cmpi ne, %[[VAL_0]], %[[CONSTANT_0]] : i32
+// CHECK:           cc.store %[[CONSTANT_1]], %[[ALLOCA_3]] : !cc.ptr<i64>
+// CHECK:           %[[VAL_0:.*]] = call @__cudaq_device_call_acquire_realtime_frame(%[[CONSTANT_3]], %[[CONSTANT_2]], %[[CONSTANT_0]], %[[CONSTANT_1]], %[[ALLOCA_0]], %[[ALLOCA_1]], %[[ALLOCA_2]]) : (i32, i32, i64, i64, !cc.ptr<!cc.ptr<i8>>, !cc.ptr<!cc.ptr<i8>>, !cc.ptr<!cc.ptr<i8>>) -> i32
+// CHECK:           %[[CMPI_0:.*]] = arith.cmpi ne, %[[VAL_0]], %[[CONSTANT_3]] : i32
 // CHECK:           cc.if(%[[CMPI_0]]) {
 // CHECK:             %[[CAST_0:.*]] = cc.cast signed %[[VAL_0]] : (i32) -> i64
 // CHECK:             func.call @__quantum__qis__trap(%[[CAST_0]]) : (i64) -> ()
@@ -30,7 +35,7 @@
 // CHECK:           %[[CAST_1:.*]] = cc.cast %[[LOAD_1]] : (!cc.ptr<i8>) -> !cc.ptr<i32>
 // CHECK:           cc.store %[[ARG0]], %[[CAST_1]] : !cc.ptr<i32>
 // CHECK:           %[[VAL_1:.*]] = call @__cudaq_device_call_dispatch_realtime_frame(%[[LOAD_0]], %[[ALLOCA_3]]) : (!cc.ptr<i8>, !cc.ptr<i64>) -> i32
-// CHECK:           %[[CMPI_1:.*]] = arith.cmpi ne, %[[VAL_1]], %[[CONSTANT_0]] : i32
+// CHECK:           %[[CMPI_1:.*]] = arith.cmpi ne, %[[VAL_1]], %[[CONSTANT_3]] : i32
 // CHECK:           cc.if(%[[CMPI_1]]) {
 // CHECK:             func.call @__cudaq_device_call_safely_release_realtime_frame(%[[LOAD_0]]) : (!cc.ptr<i8>) -> ()
 // CHECK:             %[[CAST_2:.*]] = cc.cast signed %[[VAL_1]] : (i32) -> i64
@@ -43,10 +48,7 @@
 // CHECK:         func.func private @__cudaq_device_call_acquire_realtime_frame(i32, i32, i64, i64, !cc.ptr<!cc.ptr<i8>>, !cc.ptr<!cc.ptr<i8>>, !cc.ptr<!cc.ptr<i8>>) -> i32
 // CHECK:         func.func private @__cudaq_device_call_dispatch_realtime_frame(!cc.ptr<i8>, !cc.ptr<i64>) -> i32
 // CHECK:         func.func private @__cudaq_device_call_safely_release_realtime_frame(!cc.ptr<i8>)
+// CHECK:         func.func private @llvm.memcpy.p0.p0.i64(!cc.ptr<i8>, !cc.ptr<i8>, i64, i1)
 // CHECK:         func.func private @__quantum__qis__trap(i64)
-func.func @kernel(%arg0 : i32) attributes {"cudaq-entrypoint", "cudaq-kernel", no_this} {
-  cc.device_call @fireAndForget(%arg0) : (i32) -> ()
-  return
-}
 
 func.func private @fireAndForget(i32) attributes {"cudaq-devicecall"}
